+++ Editor/ProjectGeneration/ProjectGeneration.cs
@@ -49,23 +49,12 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		const string m_SolutionProjectEntryTemplate = @"Project(""{{{0}}}"") = ""{1}"", ""{2}"", ""{{{3}}}""{4}EndProject";
 
-		readonly string m_SolutionProjectConfigurationTemplate = string.Join(k_WindowsNewline,
-			@"        {{{0}}}.Debug|Any CPU.ActiveCfg = Debug|Any CPU",
-			@"        {{{0}}}.Debug|Any CPU.Build.0 = Debug|Any CPU",
-			@"        {{{0}}}.Release|Any CPU.ActiveCfg = Release|Any CPU",
-			@"        {{{0}}}.Release|Any CPU.Build.0 = Release|Any CPU").Replace("    ", "\t");
-
-		static readonly string[] k_ReimportSyncExtensions = { ".dll", ".asmdef" };
-
-		HashSet<string> m_ProjectSupportedExtensions = new HashSet<string>();
-		HashSet<string> m_BuiltinSupportedExtensions = new HashSet<string>();
-		HashSet<string> m_DefaultSupportedExtensions = new HashSet<string>(new string[] { "dll", "asmdef", "additionalfile" });
+		HashSet<string> _supportedExtensions;
 
 		readonly string m_ProjectName;
 		internal readonly IAssemblyNameProvider m_AssemblyNameProvider;
 		readonly IFileIO m_FileIOProvider;
 		readonly IGUIDGenerator m_GUIDGenerator;
-		bool m_ShouldGenerateAll;
 		IVisualStudioInstallation m_CurrentInstallation;
 
 		public ProjectGeneration() : this(Directory.GetParent(Application.dataPath).FullName)
@@ -143,7 +132,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 					SyncProject(assembly,
 						allAssetProjectParts,
-						responseFilesData: ParseResponseFileData(assembly).ToArray());
+						ParseResponseFileData(assembly));
 				}
 
 				return true;
@@ -162,7 +151,9 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		private static bool ShouldSyncOnReimportedAsset(string asset)
 		{
-			return k_ReimportSyncExtensions.Contains(new FileInfo(asset).Extension);
+			// ".dll", ".asmdef"
+			var extension = Path.GetExtension(asset);
+			return extension == ".dll" || extension == ".asmdef";
 		}
 
 		private void RefreshCurrentInstallation()
@@ -203,8 +194,22 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		private void SetupProjectSupportedExtensions()
 		{
-			m_ProjectSupportedExtensions = new HashSet<string>(m_AssemblyNameProvider.ProjectSupportedExtensions);
-			m_BuiltinSupportedExtensions = new HashSet<string>(EditorSettings.projectGenerationBuiltinExtensions);
+			_supportedExtensions = new HashSet<string>
+			{
+				"dll",
+				"asmdef",
+				"additionalfile"
+			};
+
+			foreach (var extension in m_AssemblyNameProvider.ProjectSupportedExtensions)
+			{
+				_supportedExtensions.Add(extension);
+			}
+
+			foreach (var extension in EditorSettings.projectGenerationBuiltinExtensions)
+			{
+				_supportedExtensions.Add(extension);
+			}
 		}
 
 		private bool ShouldFileBePartOfSolution(string file)
@@ -243,20 +248,9 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		private bool IsSupportedFile(string path, out string extensionWithoutDot)
 		{
 			extensionWithoutDot = GetExtensionWithoutDot(path);
-
-			// Dll's are not scripts but still need to be included
-			if (m_DefaultSupportedExtensions.Contains(extensionWithoutDot))
-				return true;
-
-			if (m_BuiltinSupportedExtensions.Contains(extensionWithoutDot))
-				return true;
-
-			if (m_ProjectSupportedExtensions.Contains(extensionWithoutDot))
-				return true;
-
-			return false;
+			// dlls and other configured files are not scripts but still need to be included
+			return _supportedExtensions.Contains(extensionWithoutDot);
 		}
-
 
 		private static ScriptingLanguage ScriptingLanguageFor(Assembly assembly)
 		{
@@ -294,11 +288,11 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			{
 				SyncProject(assembly,
 					allAssetProjectParts,
-					responseFilesData: ParseResponseFileData(assembly).ToArray());
+					ParseResponseFileData(assembly));
 			}
 		}
 
-		private IEnumerable<ResponseFileData> ParseResponseFileData(Assembly assembly)
+		private ResponseFileData[] ParseResponseFileData(Assembly assembly)
 		{
 			var systemReferenceDirectories = CompilationPipeline.GetSystemAssemblyDirectories(assembly.compilerOptions.ApiCompatibilityLevel);
 
@@ -320,7 +314,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 					}
 			}
 
-			return responseFilesData.Select(x => x.Value);
+			return responseFilesData.Select(x => x.Value).ToArray();
 		}
 
 		private Dictionary<string, string> GenerateAllAssetProjectParts()
@@ -379,7 +373,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			if (Path.IsPathRooted(filename) && packageInfo != null)
 			{
 				// We are outside the Unity project and using a package context
-				var linkPath = SkipPathPrefix(asset.NormalizePathSeparators(), packageInfo.assetPath.NormalizePathSeparators());
+				var linkPath = XmlFilename(SkipPathPrefix(asset.NormalizePathSeparators(), packageInfo.assetPath.NormalizePathSeparators()));
 
 				builder.Append(@""">").Append(k_WindowsNewline);
 				builder.Append("      <Link>").Append(linkPath).Append("</Link>").Append(k_WindowsNewline);
@@ -394,11 +388,11 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		private void SyncProject(
 			Assembly assembly,
 			Dictionary<string, string> allAssetsProjectParts,
-			ResponseFileData[] responseFilesData)
+			ResponseFileData[] responseFileData)
 		{
 			SyncProjectFileIfNotChanged(
 				ProjectFile(assembly),
-				ProjectText(assembly, allAssetsProjectParts, responseFilesData));
+				ProjectText(assembly, allAssetsProjectParts, responseFileData));
 		}
 
 		private void SyncProjectFileIfNotChanged(string path, string newContents)
@@ -487,9 +481,9 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		private string ProjectText(Assembly assembly,
 			Dictionary<string, string> allAssetsProjectParts,
-			ResponseFileData[] responseFilesData)
+			ResponseFileData[] responseFileData)
 		{
-			ProjectHeader(assembly, responseFilesData, out StringBuilder projectBuilder);
+			ProjectHeader(assembly, responseFileData, out StringBuilder projectBuilder);
 
 			var references = new List<string>();
 
@@ -524,7 +518,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 			projectBuilder.Append(@"  <ItemGroup>").Append(k_WindowsNewline);
 
-			var responseRefs = responseFilesData.SelectMany(x => x.FullPathReferences.Select(r => r));
+			var responseRefs = responseFileData.SelectMany(x => x.FullPathReferences.Select(r => r));
 			var internalAssemblyReferences = assembly.assemblyReferences
 				.Where(i => !i.sourceFiles.Any(ShouldFileBePartOfSolution)).Select(i => i.outputPath);
 			var allReferences =
@@ -598,11 +592,20 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		public string SolutionFile()
 		{
+			return SolutionFileImpl();
+		}
+
+		internal virtual string SolutionFileImpl()
+		{
 			return Path.Combine(ProjectDirectory.NormalizePathSeparators(), $"{InvalidCharactersRegexPattern.Replace(m_ProjectName, "_")}.sln");
 		}
 
-		internal string GetLangVersion(Assembly assembly)
+		internal string GetLangVersion(Assembly assembly, ResponseFileData[] responseFileData)
 		{
+			var langVersion = GetOtherArguments(responseFileData, "langversion").FirstOrDefault();
+			if (!string.IsNullOrEmpty(langVersion))
+				return langVersion;
+
 			var targetLanguageVersion = "latest"; // danger: latest is not the same absolute value depending on the VS version.
 			if (m_CurrentInstallation != null)
 			{
@@ -616,9 +619,9 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			return targetLanguageVersion;
 		}
 
-		private static IEnumerable<string> GetOtherArguments(ResponseFileData[] responseFilesData, HashSet<string> names)
+		private static IEnumerable<string> GetOtherArguments(ResponseFileData[] responseFileData, string name)
 		{
-			var lines = responseFilesData
+			var lines = responseFileData
 				.SelectMany(x => x.OtherArguments)
 				.Where(l => !string.IsNullOrEmpty(l))
 				.Select(l => l.Trim())
@@ -634,7 +637,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 					.Substring(1, index - 1)
 					.Trim();
 				
-				if (!names.Contains(key))
+				if (name != key)
 					continue;
 
 				if (argument.Length <= index)
@@ -646,7 +649,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			}
 		}
 
-		private void SetAnalyzerAndSourceGeneratorProperties(Assembly assembly, ResponseFileData[] responseFilesData, ProjectProperties properties)
+		private void SetAnalyzerAndSourceGeneratorProperties(Assembly assembly, ResponseFileData[] responseFileData, ProjectProperties properties)
 		{
 			if (m_CurrentInstallation == null || !m_CurrentInstallation.SupportsAnalyzers)
 				return;
@@ -678,8 +681,9 @@ namespace Microsoft.Unity.VisualStudio.Editor
 #endif
 
 			// Analyzers and additional files provided by csc.rsp
-			analyzers.AddRange(GetOtherArguments(responseFilesData, new HashSet<string>(new[] { "analyzer", "a" })));
-			additionalFilePaths.AddRange(GetOtherArguments(responseFilesData, new HashSet<string>(new[] { "additionalfile" })));
+			analyzers.AddRange(GetOtherArguments(responseFileData, "analyzer"));
+			analyzers.AddRange(GetOtherArguments(responseFileData, "a"));
+			additionalFilePaths.AddRange(GetOtherArguments(responseFileData, "additionalfile"));
 
 			properties.RulesetPath = ToNormalizedPath(rulesetPath);
 			properties.Analyzers = ToNormalizedPaths(analyzers);
@@ -705,7 +709,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		private void ProjectHeader(
 			Assembly assembly,
-			ResponseFileData[] responseFilesData,
+			ResponseFileData[] responseFileData,
 			out StringBuilder headerBuilder
 		)
 		{
@@ -714,13 +718,13 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			var projectProperties = new ProjectProperties
 			{
 				ProjectGuid = ProjectGuid(assembly),
-				LangVersion = GetLangVersion(assembly),
+				LangVersion = GetLangVersion(assembly, responseFileData),
 				AssemblyName = assembly.name,
 				RootNamespace = GetRootNamespace(assembly),
 				OutputPath = assembly.outputPath,
 				// RSP alterable
-				Defines = assembly.defines.Concat(responseFilesData.SelectMany(x => x.Defines)).Distinct().ToArray(),
-				Unsafe = assembly.compilerOptions.AllowUnsafeCode | responseFilesData.Any(x => x.Unsafe),
+				Defines = assembly.defines.Concat(responseFileData.SelectMany(x => x.Defines)).Distinct().ToArray(),
+				Unsafe = assembly.compilerOptions.AllowUnsafeCode | responseFileData.Any(x => x.Unsafe),
 				// VSTU Flavoring
 				FlavoringProjectType = projectType + ":" + (int)projectType,
 				FlavoringBuildTarget = EditorUserBuildSettings.activeBuildTarget + ":" + (int)EditorUserBuildSettings.activeBuildTarget,
@@ -728,7 +732,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 				FlavoringPackageVersion = VisualStudioIntegration.PackageVersion(),
 			};
 
-			SetAnalyzerAndSourceGeneratorProperties(assembly, responseFilesData, projectProperties);
+			SetAnalyzerAndSourceGeneratorProperties(assembly, responseFileData, projectProperties);
 
 			GetProjectHeader(projectProperties, out headerBuilder);
 		}
@@ -832,27 +836,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		{
 		}
 
-		private static string GetSolutionText()
-		{
-			return string.Join(k_WindowsNewline,
-			@"",
-			@"Microsoft Visual Studio Solution File, Format Version {0}",
-			@"# Visual Studio {1}",
-			@"{2}",
-			@"Global",
-			@"    GlobalSection(SolutionConfigurationPlatforms) = preSolution",
-			@"        Debug|Any CPU = Debug|Any CPU",
-			@"        Release|Any CPU = Release|Any CPU",
-			@"    EndGlobalSection",
-			@"    GlobalSection(ProjectConfigurationPlatforms) = postSolution",
-			@"{3}",
-			@"    EndGlobalSection",
-			@"{4}",
-			@"EndGlobal",
-			@"").Replace("    ", "\t");
-		}
-
-		private void SyncSolution(IEnumerable<Assembly> assemblies)
+		internal virtual void SyncSolution(IEnumerable<Assembly> assemblies)
 		{
 			if (InvalidCharactersRegexPattern.IsMatch(ProjectDirectory))
 				Debug.LogWarning("Project path contains special characters, which can be an issue when opening Visual Studio");
@@ -862,15 +846,45 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			SyncSolutionFileIfNotChanged(solutionFile, SolutionText(assemblies, previousSolution));
 		}
 
-		private string SolutionText(IEnumerable<Assembly> assemblies, Solution previousSolution = null)
+		internal virtual string SolutionText(IEnumerable<Assembly> assemblies, Solution previousSolution = null)
 		{
 			const string fileversion = "12.00";
 			const string vsversion = "15";
 
+			var projects = GetSolutionProjects(assemblies, previousSolution);
+			var properties = previousSolution != null ? previousSolution.Properties : null;
+
+			string propertiesText = GetPropertiesText(properties);
+			string projectEntriesText = GetProjectEntriesText(projects);
+
+			// do not generate configurations for SolutionFolders
+			var configurableProjects = projects.Where(p => !p.IsSolutionFolderProjectFactory());
+			string projectConfigurationsText = string.Join(k_WindowsNewline, configurableProjects.Select(p => GetProjectActiveConfigurations(p.ProjectGuid)).ToArray());
+
+			const string solutionText =
+				"" + k_WindowsNewline
+				+ "Microsoft Visual Studio Solution File, Format Version {0}" + k_WindowsNewline
+				+ "# Visual Studio {1}" + k_WindowsNewline
+				+ "{2}" + k_WindowsNewline
+				+ "Global" + k_WindowsNewline
+				+ "\tGlobalSection(SolutionConfigurationPlatforms) = preSolution" + k_WindowsNewline
+				+ "\t\tDebug|Any CPU = Debug|Any CPU" + k_WindowsNewline
+				+ "\t\tRelease|Any CPU = Release|Any CPU" + k_WindowsNewline
+				+ "\tEndGlobalSection" + k_WindowsNewline
+				+ "\tGlobalSection(ProjectConfigurationPlatforms) = postSolution" + k_WindowsNewline
+				+ "{3}" + k_WindowsNewline
+				+ "\tEndGlobalSection" + k_WindowsNewline
+				+ "{4}" + k_WindowsNewline
+				+ "EndGlobal" + k_WindowsNewline
+				+ "";
+
+			return string.Format(solutionText, fileversion, vsversion, projectEntriesText, projectConfigurationsText, propertiesText);
+		}
+
+		internal List<SolutionProjectEntry> GetSolutionProjects(IEnumerable<Assembly> assemblies, Solution previousSolution = null)
+		{
 			var relevantAssemblies = RelevantAssembliesForMode(assemblies);
 			var generatedProjects = ToProjectEntries(relevantAssemblies).ToList();
-
-			SolutionProperties[] properties = null;
 
 			// First, add all projects generated by Unity to the solution
 			var projects = new List<SolutionProjectEntry>();
@@ -884,17 +898,9 @@ namespace Microsoft.Unity.VisualStudio.Editor
 					.Where(p => generatedProjects.All(gp => gp.FileName != p.FileName));
 
 				projects.AddRange(externalProjects);
-				properties = previousSolution.Properties;
 			}
 
-			string propertiesText = GetPropertiesText(properties);
-			string projectEntriesText = GetProjectEntriesText(projects);
-
-			// do not generate configurations for SolutionFolders
-			var configurableProjects = projects.Where(p => !p.IsSolutionFolderProjectFactory());
-			string projectConfigurationsText = string.Join(k_WindowsNewline, configurableProjects.Select(p => GetProjectActiveConfigurations(p.ProjectGuid)).ToArray());
-
-			return string.Format(GetSolutionText(), fileversion, vsversion, projectEntriesText, projectConfigurationsText, propertiesText);
+			return projects;
 		}
 
 		private static IEnumerable<Assembly> RelevantAssembliesForMode(IEnumerable<Assembly> assemblies)
@@ -956,6 +962,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		private IEnumerable<SolutionProjectEntry> ToProjectEntries(IEnumerable<Assembly> assemblies)
 		{
 			foreach (var assembly in assemblies)
+			{
 				yield return new SolutionProjectEntry()
 				{
 					ProjectFactoryGuid = SolutionGuid(assembly),
@@ -964,6 +971,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 					ProjectGuid = ProjectGuid(assembly),
 					Metadata = k_WindowsNewline
 				};
+			}
 		}
 
 		/// <summary>
@@ -971,8 +979,14 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		/// </summary>
 		private string GetProjectActiveConfigurations(string projectGuid)
 		{
+			const string solutionProjectConfigurationTemplate =
+				"\t\t{{{0}}}.Debug|Any CPU.ActiveCfg = Debug|Any CPU" + k_WindowsNewline
+				+ "\t\t{{{0}}}.Debug|Any CPU.Build.0 = Debug|Any CPU" + k_WindowsNewline
+				+ "\t\t{{{0}}}.Release|Any CPU.ActiveCfg = Release|Any CPU" + k_WindowsNewline
+				+ "\t\t{{{0}}}.Release|Any CPU.Build.0 = Release|Any CPU";
+
 			return string.Format(
-				m_SolutionProjectConfigurationTemplate,
+				solutionProjectConfigurationTemplate,
 				projectGuid);
 		}
 
@@ -987,7 +1001,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			{
 				// We have to normalize the path, because the PackageManagerRemapper assumes
 				// dir seperators will be os specific.
-				var absolutePath = Path.GetFullPath(path.NormalizePathSeparators());
+				var absolutePath = FileUtility.GetAbsolutePath(path.NormalizePathSeparators());
 				path = SkipPathPrefix(absolutePath, projectDir);
 			}