+++ Editor/VisualStudioEditor.cs
@@ -68,7 +68,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			{
 				return Discovery
 					.GetVisualStudioInstallations()
-					.ToDictionary(i => Path.GetFullPath(i.Path), i => i);
+					.ToDictionary(i => FileUtility.GetAbsolutePath(i.Path), i => i);
 			}
 			catch (Exception ex)
 			{
@@ -97,7 +97,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		internal virtual bool TryGetVisualStudioInstallationForPath(string editorPath, bool lookupDiscoveredInstallations, out IVisualStudioInstallation installation)
 		{
-			editorPath = Path.GetFullPath(editorPath);
+			editorPath = FileUtility.GetAbsolutePath(editorPath);
 
 			// lookup for well known installations
 			if (lookupDiscoveredInstallations && _discoverInstallations.Result.TryGetValue(editorPath, out installation))