+++ CHANGELOG.md
@@ -1,5 +1,46 @@
 # Code Editor Package for Visual Studio
 
+## [2.0.27] - 2026-01-20
+
+Integration:
+
+- Fix Visual Studio Integration to properly wait for the solution to be opened.
+
+Project generation:
+
+- Disable Workspace-based development feature in `settings.json`.
+- Ensure that we only have one `sln` or `slnx` file at a time.
+- Properly handle filenames with special characters in `link` tags.
+
+## [2.0.26] - 2025-12-16
+
+Integration:
+
+- Fix handling of asset-pipeline refresh-mode setting.
+
+Project generation:
+
+- Add `EnableOnDemandExcludedFolderLoading` capability when generating SDK-Style project.
+
+## [2.0.25] - 2025-09-16
+
+Internal:
+
+- Fixes for release validation and release process.
+
+## [2.0.24] - 2025-09-04
+
+Integration:
+
+- Remove support for `Visual Studio for Mac`. Please use `VS Code` going forward.
+- Performance optimizations.
+
+Project generation:
+
+- Allow customization of `langversion` when using a `rsp` file.
+- Move to `slnx` solution generation when using `SDK-Style` projects.
+- Both `VS Code` and `Visual Studio 2026` are now using `SDK-Style` projects by default.
+
 ## [2.0.23] - 2025-02-03
 
 Integration:
@@ -12,7 +53,6 @@ Project generation:
 - Identify `asset`, `meta`, `prefab` and `unity` files as `yaml` (Visual Studio Code).
 - Add `sln`/`csproj` file nesting (Visual Studio Code).
 - Improve SDK style project generation.
-
 
 ## [2.0.22] - 2023-10-03